feat: moves data fetching logic to myForestContextV2
