+++ src/features/common/Layout/MyForestContextV2.tsx
@@ -17,6 +17,7 @@ import {
   MyContributionsSingleProject,
   MapLocation,
   MyContributionsMapItem,
+  Leaderboard,
 } from '../types/myForestv2';
 import { updateStateWithTrpcData } from '../../../utils/trpcHelpers';
 import { SetState } from '../types/common';
@@ -46,6 +47,8 @@ interface UserInfo {
 interface MyForestContextV2Interface {
   projectListResult: ProjectListResponse | undefined;
   contributionsResult: ContributionsResponse | undefined;
+  leaderboardResult: Leaderboard | undefined;
+  isLeaderboardLoaded: boolean;
   contributionsMap: Map<string, MyContributionsMapItem> | undefined;
   registrationGeojson: RegistrationGeojson[];
   donationGeojson: DonationGeojson[];
@@ -62,6 +65,10 @@ export const MyForestProviderV2: FC = ({ children }) => {
     useState<ProjectListResponse>();
   const [contributionsResult, setContributionsResult] =
     useState<ContributionsResponse>();
+  const [leaderboardResult, setLeaderboardResult] = useState<
+    Leaderboard | undefined
+  >();
+  const [isLeaderboardLoaded, setIsLeaderboardLoaded] = useState(false);
   const [contributionsMap, setContributionsMap] =
     useState<Map<string, MyContributionsMapItem>>();
   const [registrationGeojson, setRegistrationGeojson] = useState<
@@ -73,6 +80,10 @@ export const MyForestProviderV2: FC = ({ children }) => {
 
   const _projectList = trpc.myForestV2.projectList.useQuery();
   const _contributions = trpc.myForestV2.contributions.useQuery({
+    profileId: `${userInfo?.profileId}`,
+    slug: `${userInfo?.slug}`,
+  });
+  const _leaderboard = trpc.myForestV2.leaderboard.useQuery({
     profileId: `${userInfo?.profileId}`,
     slug: `${userInfo?.slug}`,
   });
@@ -92,6 +103,13 @@ export const MyForestProviderV2: FC = ({ children }) => {
       );
     }
   }, [_contributions.data]);
+
+  useEffect(() => {
+    if (_leaderboard.data) {
+      updateStateWithTrpcData(_leaderboard, setLeaderboardResult, setErrors);
+      setIsLeaderboardLoaded(true);
+    }
+  }, [_leaderboard?.data]);
 
   //format geojson
   const _generateDonationGeojson = (
@@ -160,6 +178,8 @@ export const MyForestProviderV2: FC = ({ children }) => {
     () => ({
       projectListResult,
       contributionsResult,
+      leaderboardResult,
+      isLeaderboardLoaded,
       contributionsMap,
       registrationGeojson,
       donationGeojson,
@@ -169,6 +189,8 @@ export const MyForestProviderV2: FC = ({ children }) => {
     [
       projectListResult,
       contributionsResult,
+      leaderboardResult,
+      isLeaderboardLoaded,
       contributionsMap,
       registrationGeojson,
       donationGeojson,
